% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_session.R
\name{attachMriSession}
\alias{attachMriSession}
\alias{newSessionFromDirectory}
\title{Attach a session directory}
\usage{
attachMriSession(directory)

newSessionFromDirectory(directory)
}
\arguments{
\item{directory}{A string giving the path to the session directory (i.e.
the parent of the \code{tractor} subdirectory in which TractoR-managed files
live).}
}
\value{
An \link{MriSession} object.
}
\description{
These functions create an \link{MriSession} object representing an existing
session directory on disk.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
