% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{coregisterDataVolumesForSession}
\alias{coregisterDataVolumesForSession}
\title{Coregister the volumes of a 4D data set within a session}
\usage{
coregisterDataVolumesForSession(session, type, reference = 1,
  useMask = FALSE, nLevels = 2, method = c("niftyreg", "fsl", "none"),
  options = list(), ...)
}
\arguments{
\item{session}{An \link{MriSession} object.}

\item{type}{A string giving the (session-relative) type of the 4D data,
e.g. \code{"diffusion"} or \code{"functional"}.}

\item{reference}{An integer index into the fourth dimension of the raw
data, giving the volume to register to, or an
\link[tractor.base:MriImage]{tractor.base::MriImage} to use as the target directly.}

\item{useMask}{Boolean value: should the session's mask for \code{type} be used
to weight or restrict the registration?}

\item{nLevels}{An integer giving the number of resolution levels to use for
the registration.}

\item{method}{A string, one of \code{"niftyreg"}, \code{"fsl"} or \code{"none"}. With
\code{"none"}, an identity transform is stored and the raw data is simply
copied to the coregistered data location.}

\item{options}{A list of additional method-specific linear registration
options.}

\item{...}{Additional arguments to \code{\link[tractor.reg:registerImages]{tractor.reg::registerImages()}}.}
}
\value{
The resulting \link[tractor.reg:Registration]{tractor.reg::Registration} object.
}
\description{
This function coregisters each volume of a session's raw 4D data (of the
specified \code{type}, e.g. \code{"diffusion"} or \code{"functional"}) to a common
reference volume, and writes the resulting registered data set to file.
This is typically used for motion correction of functional or diffusion
time series that are not otherwise handled by a dedicated tool such as
FSL's \code{eddy} (see \code{\link[=runEddyWithSession]{runEddyWithSession()}}).
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
