% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{flipGradientVectorsForSession}
\alias{flipGradientVectorsForSession}
\title{Flip gradient direction components for a session}
\usage{
flipGradientVectorsForSession(session, axes, unrotated = FALSE)
}
\arguments{
\item{session}{An \link{MriSession} object, which must already have a diffusion
scheme associated with it.}

\item{axes}{An integer vector giving the gradient direction component(s)
to flip: some subset of 1 (x), 2 (y) and 3 (z).}

\item{unrotated}{Boolean value: should the unrotated (as opposed to
eddy-current-corrected) gradient directions be flipped?}
}
\value{
This function is called for its side effect.
}
\description{
This function negates one or more components of a session's diffusion
gradient directions, and writes the updated scheme back to file. This is
typically needed to correct for a sign convention mismatch between the
scanner and the tractography software being used.
}
\author{
Jon Clayden
}
