% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl.R
\name{getBedpostNumberOfFibresForSession}
\alias{getBedpostNumberOfFibresForSession}
\title{Count the fibre orientations modelled by bedpostx for a session}
\usage{
getBedpostNumberOfFibresForSession(session)
}
\arguments{
\item{session}{An \link{MriSession} object.}
}
\value{
The number of fibre orientations modelled, which may be zero if
BEDPOSTX has not been run for the session.
}
\description{
This function counts the number of fibre orientation (\code{"avf"}, anisotropic
volume fraction) maps produced by a previous run of FSL's \code{bedpostx} tool
for a session (see \code{\link[=runBedpostWithSession]{runBedpostWithSession()}}).
}
\author{
Jon Clayden
}
