% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl.R
\name{getFslVersion}
\alias{getFslVersion}
\title{Determine the installed version of FSL}
\usage{
getFslVersion()
}
\value{
An integer encoding the FSL version as \code{10000*major + 100*minor + patch}, or \code{NULL} if the version cannot be determined (e.g. because FSL
does not appear to be installed).
}
\description{
This function determines the version of FSL installed on the system, if
any, by reading the \code{etc/fslversion} file within the directory pointed to
by the \code{FSLDIR} environment variable.
}
\author{
Jon Clayden
}
