% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_session.R
\name{getImageCountForSession}
\alias{getImageCountForSession}
\title{Count the images of a particular type associated with a session}
\usage{
getImageCountForSession(session, type, place = NULL)
}
\arguments{
\item{session}{An \link{MriSession} object.}

\item{type}{A string giving the (session-relative) type of image required.}

\item{place}{A string giving the subdirectory in which the image should be
found, or \code{NULL} to look in the type's standard location.}
}
\value{
The number of images of the specified type found, which may be
zero.
}
\description{
This function counts the number of images of a particular type associated
with a session, using the \code{index} mechanism supported by some image types
(such as per-fibre \code{"avf"} maps produced by FSL-BEDPOSTX). It works by
testing sequentially-indexed images for existence until one is not found.
}
\author{
Jon Clayden
}
