% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{getFileNameForStandardImage}
\alias{getFileNameForStandardImage}
\alias{getStandardImage}
\title{Obtain a standard-space (MNI) reference image}
\usage{
getFileNameForStandardImage(name = c("brain", "face"),
  errorIfMissing = TRUE)

getStandardImage(name, errorIfMissing = TRUE, ...)
}
\arguments{
\item{name}{For \code{getFileNameForStandardImage()}, one of \code{"brain"} or
\code{"face"}. For \code{getStandardImage()}, any type name understood by
\code{getFileNameForStandardImage()}.}

\item{errorIfMissing}{Boolean value: if \code{TRUE}, an error is raised if the
standard images cannot be found; otherwise \code{NULL} is returned.}

\item{...}{Additional arguments to \code{\link[tractor.base:readImageFile]{tractor.base::readImageFile()}}.}
}
\value{
For \code{getFileNameForStandardImage()}, a string giving the path to
the relevant image, or \code{NULL}. For \code{getStandardImage()}, the
corresponding \link[tractor.base:MriImage]{tractor.base::MriImage} object, or \code{NULL}.
}
\description{
These functions return the path to, or contents of, one of the standard
MNI-space reference images shipped with TractoR (under
\code{share/tractor/mni}), such as the standard brain volume used as a
registration target for the \code{"mni"} space (see \link{MriSession}).
}
\author{
Jon Clayden
}
