% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{getVolumeTransformationForSession}
\alias{getVolumeTransformationForSession}
\title{Read the volume-to-volume transformation for a session}
\usage{
getVolumeTransformationForSession(session, type)
}
\arguments{
\item{session}{An \link{MriSession} object.}

\item{type}{A string giving the (session-relative) type of the 4D data,
e.g. \code{"diffusion"} or \code{"functional"}.}
}
\value{
The relevant \link[tractor.reg:Registration]{tractor.reg::Registration} object.
}
\description{
This function reads the registration relating the raw volumes of a
session's 4D data of a particular \code{type} to their common reference volume,
as produced by \code{\link[=coregisterDataVolumesForSession]{coregisterDataVolumesForSession()}} or, for diffusion data,
by eddy current correction (see \code{\link[=readEddyCorrectTransformsForSession]{readEddyCorrectTransformsForSession()}}).
}
\author{
Jon Clayden
}
