% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{guessSpace}
\alias{guessSpace}
\title{Guess the space associated with an image}
\usage{
guessSpace(image, session = NULL, errorIfOutOfSession = TRUE)
}
\arguments{
\item{image}{An \link[tractor.base:MriImage]{tractor.base::MriImage} object, or a string giving the
path to an image file. Internal (in-memory only) images are not
associated with any space.}

\item{session}{An \link{MriSession} object, or \code{NULL} to try to infer the
session from the image's path (which will only work if the image lies
within a session directory).}

\item{errorIfOutOfSession}{Boolean value: if \code{TRUE}, an error is raised if
\code{session} is \code{NULL} and cannot be inferred from the image's path;
otherwise \code{NULL} is returned in that case.}
}
\value{
A string naming the space, in the form understood by
\code{\link[=transformImageToSpace]{transformImageToSpace()}} and related functions, or \code{NULL} if the space
cannot be determined.
}
\description{
This function attempts to work out which named space (such as \code{"mni"},
or a session-relative space such as \code{"diffusion"}) an image belongs to, by
comparing the directory containing the image to the standard locations
associated with a session, or to the location of TractoR's standard-space
reference images.
}
\author{
Jon Clayden
}
