% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl.R
\name{runBedpostWithSession}
\alias{runBedpostWithSession}
\title{Run FSL's bedpostx tool for a session}
\usage{
runBedpostWithSession(session, nFibres = 3)
}
\arguments{
\item{session}{An \link{MriSession} object, or a string giving the path to a
session directory.}

\item{nFibres}{An integer giving the maximum number of fibre orientations
to model per voxel.}
}
\value{
This function is called for its side effect.
}
\description{
This function runs FSL's \code{bedpostx} tool to fit a ball-and-sticks
diffusion model to a session's preprocessed diffusion-weighted data, via
the \code{bedpostx} TractoR workflow (see \code{\link[=runWorkflow]{runWorkflow()}}). Any existing
BEDPOSTX output for the session is removed first. A multi-shell model is
used automatically if the session's diffusion scheme has more than one
nonzero b-value.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\seealso{
\code{\link[=getBedpostNumberOfFibresForSession]{getBedpostNumberOfFibresForSession()}}, which counts the number
of fibre orientation maps produced by a previous run.
}
\author{
Jon Clayden
}
