% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl.R
\name{runBetWithSession}
\alias{runBetWithSession}
\title{Run FSL's bet tool for a session}
\usage{
runBetWithSession(session, intensityThreshold = 0.5, verticalGradient = 0)
}
\arguments{
\item{session}{An \link{MriSession} object.}

\item{intensityThreshold}{A numeric fractional intensity threshold, between
0 and 1, smaller values giving larger brain outline estimates. Passed to
\code{bet}'s \code{-f} option.}

\item{verticalGradient}{A numeric threshold gradient, applied vertically,
in the range -1 to 1. Positive values give larger brain outlines at the
bottom and smaller ones at the top of the head. Passed to \code{bet}'s \code{-g}
option.}
}
\value{
This function is called for its side effect.
}
\description{
This function runs FSL's \code{bet} (Brain Extraction Tool) on a session's
reference diffusion b=0 volume, via the \code{bet-diffusion} TractoR workflow
(see \code{\link[=runWorkflow]{runWorkflow()}}).
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
