% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl.R
\name{runDtifitWithSession}
\alias{runDtifitWithSession}
\title{Run FSL's dtifit tool for a session}
\usage{
runDtifitWithSession(session, weightedLeastSquares = FALSE)
}
\arguments{
\item{session}{An \link{MriSession} object.}

\item{weightedLeastSquares}{Boolean value: should tensors be fitted with
weighted, rather than ordinary, least-squares?}
}
\value{
This function is called for its side effect.
}
\description{
This function runs FSL's \code{dtifit} tool to fit diffusion tensors to a
session's preprocessed diffusion-weighted data, via the \code{dtifit} TractoR
workflow (see \code{\link[=runWorkflow]{runWorkflow()}}).
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\seealso{
\code{\link[=estimateDiffusionTensors]{estimateDiffusionTensors()}} and
\code{\link[=createDiffusionTensorImagesForSession]{createDiffusionTensorImagesForSession()}}, which provide an alternative,
pure-R tensor fitting pipeline.
}
\author{
Jon Clayden
}
