% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl.R
\name{runEddyCorrectWithSession}
\alias{runEddyCorrectWithSession}
\title{Run FSL's eddy_correct tool for a session}
\usage{
runEddyCorrectWithSession(session, refVolume)
}
\arguments{
\item{session}{An \link{MriSession} object, or a string giving the path to a
session directory.}

\item{refVolume}{An integer giving the index of the volume to use as the
registration target.}
}
\value{
The resulting \link[tractor.reg:Registration]{tractor.reg::Registration} object, invisibly.
}
\description{
This function runs FSL's older, simpler \code{eddy_correct} tool to correct a
session's diffusion-weighted data for eddy-current-induced distortion, by
registering each volume to a single reference volume, via the
\code{eddycorrect} TractoR workflow (see \code{\link[=runWorkflow]{runWorkflow()}}). The resulting affine
transforms are read back into the session on completion (see
\code{\link[=readEddyCorrectTransformsForSession]{readEddyCorrectTransformsForSession()}}).
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\seealso{
\code{\link[=runEddyWithSession]{runEddyWithSession()}}, which uses the more modern \code{eddy} tool
instead.
}
\author{
Jon Clayden
}
