% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freesurfer.R
\name{runFreesurferForSession}
\alias{runFreesurferForSession}
\title{Run FreeSurfer's recon-all pipeline for a session}
\usage{
runFreesurferForSession(session, options = NULL)
}
\arguments{
\item{session}{An \link{MriSession} object.}

\item{options}{A string, or character vector, of additional command-line
options to pass to \code{recon-all}.}
}
\value{
This function is called for its side effect.
}
\description{
This function runs FreeSurfer's \code{recon-all} cortical reconstruction
pipeline for a session's structural image, via the \code{reconall} TractoR
workflow (see \code{\link[=runWorkflow]{runWorkflow()}}). FreeSurfer must be installed and set up
appropriately for this function to work. Any existing FreeSurfer output
for the session is removed first.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
