% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{runWorkflow}
\alias{runWorkflow}
\title{Run a TractoR workflow}
\usage{
runWorkflow(name, session, ..., .args = "")
}
\arguments{
\item{name}{A string giving the name of the workflow to run (see
\code{\link[=findWorkflow]{findWorkflow()}}).}

\item{session}{An \link{MriSession} object, or a string giving the path to a
session directory, on which the workflow will operate.}

\item{...}{Named configuration variables to make available to the
workflow script as environment variables.}

\item{.args}{A string, or character vector, of additional command-line
arguments to make available to the workflow script (as
\code{TRACTOR_COMMAND_ARGS}).}
}
\value{
The workflow's exit code, invisibly. An error is raised if this is
nonzero.
}
\description{
This function runs a TractoR "workflow": a shell script, usually wrapping
one or more calls to third-party command-line tools such as FSL or
FreeSurfer programs, which together implement a step of a typical
tractography or other MRI processing pipeline. The workflow's location is
found using \code{\link[=findWorkflow]{findWorkflow()}} and its prerequisites checked using
\code{\link[=precheckWorkflow]{precheckWorkflow()}} before it is run, with a suitable environment
including session-specific and other configuration variables.
}
\references{
Please cite the following reference when using TractoR in your
work:

J.D. Clayden, S. Muñoz Maniega, A.J. Storkey, M.D. King, M.E. Bastin & C.A.
Clark (2011). TractoR: Magnetic resonance imaging and tractography with R.
Journal of Statistical Software 44(8):1-18. \doi{10.18637/jss.v044.i08}.
}
\author{
Jon Clayden
}
