% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freesurfer.R, R/fsl.R, R/mrtrix.R
\name{showImagesInFreeview}
\alias{showImagesInFreeview}
\alias{showImagesInFsleyes}
\alias{showImagesInFslview}
\alias{showImagesInMrview}
\title{Display images in an external viewer program}
\usage{
showImagesInFreeview(imageFileNames, wait = FALSE, lookupTable = NULL,
  opacity = NULL)

showImagesInFsleyes(imageFileNames, wait = FALSE, lookupTable = NULL,
  opacity = NULL)

showImagesInFslview(imageFileNames, wait = FALSE, lookupTable = NULL,
  opacity = NULL)

showImagesInMrview(imageFileNames, wait = FALSE, lookupTable = NULL,
  opacity = NULL)
}
\arguments{
\item{imageFileNames}{A character vector of paths to image files to
display. The first is taken to be the main image, and any others are
shown as overlays on top of it.}

\item{wait}{Boolean value: should the function wait for the viewer program
to exit before returning?}

\item{lookupTable}{An optional character vector of colour lookup table (or
colour map) names, of the same length as \code{imageFileNames}, or which will
be recycled to that length.}

\item{opacity}{An optional numeric vector of opacity values, in the range
0 to 1, of the same length as \code{imageFileNames}, or which will be
recycled to that length.}
}
\value{
The image file names, invisibly.
}
\description{
These functions display one or more images by launching an external image
viewer program as a subprocess: \code{showImagesInFreeview()} for FreeSurfer's
\code{freeview}, \code{showImagesInFsleyes()} for FSL's \code{fsleyes},
\code{showImagesInFslview()} for FSL's (deprecated) \code{fslview}, and
\code{showImagesInMrview()} for MRtrix's \code{mrview}. The relevant program must be
installed and on the system's path. These functions are not usually called
directly by users; instead see \code{\link[=showImagesInViewer]{showImagesInViewer()}}, which also supports
TractoR's own internal image viewer and handles image format conversion
where necessary.
}
\seealso{
\code{\link[=showImagesInViewer]{showImagesInViewer()}}, which provides a unified, higher-level
interface to these and TractoR's internal viewer.
}
\author{
Jon Clayden
}
